feat:[SSCA-965]: Added pipeline name in artifact details and refactored pipeline code
