+++ ssca-manager/service/src/main/java/io/harness/ssca/services/ArtifactServiceImpl.java
@@ -44,6 +44,7 @@ import io.harness.ssca.entities.EnforcementSummaryEntity;
 import io.harness.ssca.entities.EnforcementSummaryEntity.EnforcementSummaryEntityKeys;
 import io.harness.ssca.utils.SBOMUtils;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.google.inject.Inject;
 import java.net.URI;
 import java.time.Instant;
@@ -83,6 +84,7 @@ public class ArtifactServiceImpl implements ArtifactService {
   @Inject EnforcementSummaryService enforcementSummaryService;
   @Inject NormalisedSbomComponentService normalisedSbomComponentService;
   @Inject CdInstanceSummaryService cdInstanceSummaryService;
+  @Inject PipelineService pipelineService;
 
   private final String GCP_REGISTRY_HOST = "grc.io";
 
@@ -180,6 +182,7 @@ public class ArtifactServiceImpl implements ArtifactService {
       throw new NotFoundException(
           String.format("Artifact with artifactId [%s] and tag [%s] is not found", artifactId, tag));
     }
+    JsonNode rootNode = pipelineService.getPmsExecutionSummary(artifact);
     return new ArtifactDetailResponse()
         .id(artifact.getArtifactId())
         .name(artifact.getName())
@@ -190,6 +193,7 @@ public class ArtifactServiceImpl implements ArtifactService {
         .prodEnvCount(artifact.getProdEnvCount().intValue())
         .nonProdEnvCount(artifact.getNonProdEnvCount().intValue())
         .buildPipelineId(artifact.getPipelineId())
+        .buildPipelineName(pipelineService.getPipelineName(rootNode))
         .buildPipelineExecutionId(artifact.getPipelineExecutionId())
         .orchestrationId(artifact.getOrchestrationId());
   }
