+++ ssca-manager/service/src/main/java/io/harness/ssca/services/CdInstanceSummaryServiceImpl.java
@@ -9,15 +9,10 @@ package io.harness.ssca.services;
 
 import io.harness.entities.Instance;
 import io.harness.exception.InvalidArgumentsException;
-import io.harness.pipeline.remote.PipelineServiceClient;
-import io.harness.remote.client.NGRestUtils;
 import io.harness.repositories.CdInstanceSummaryRepo;
 import io.harness.repositories.EnforcementSummaryRepo;
-import io.harness.serializer.JsonUtils;
 import io.harness.spec.server.ssca.v1.model.ArtifactDeploymentViewRequestBody;
 import io.harness.ssca.beans.EnvType;
-import io.harness.ssca.beans.SLSAVerificationSummary;
-import io.harness.ssca.beans.SLSAVerificationSummary.SLSAVerificationSummaryBuilder;
 import io.harness.ssca.entities.ArtifactEntity;
 import io.harness.ssca.entities.CdInstanceSummary;
 import io.harness.ssca.entities.CdInstanceSummary.CdInstanceSummaryKeys;
@@ -28,7 +23,6 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.inject.Inject;
 import java.time.Instant;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
 import java.util.regex.Pattern;
@@ -43,30 +37,7 @@ public class CdInstanceSummaryServiceImpl implements CdInstanceSummaryService {
   @Inject CdInstanceSummaryRepo cdInstanceSummaryRepo;
   @Inject ArtifactService artifactService;
   @Inject EnforcementSummaryRepo enforcementSummaryRepo;
-  @Inject PipelineServiceClient pipelineServiceClient;
-
-  private static String IDENTIFIER = "identifier";
-  private static String STEP_TYPE = "stepType";
-
-  private static String EXECUTION_GRAPH = "executionGraph";
-  private static String SLSA_VERIFICATION_STEP_ID = "SlsaVerification";
-  private static String NODE_MAP = "nodeMap";
-  private static String OUTCOMES = "outcomes";
-
-  private static String PIPELINE_EXECUTION_SUMMARY = "pipelineExecutionSummary";
-  private static String PIPELINE_NAME = "name";
-  private static String RUN_SEQUENCE = "runSequence";
-  private static String EXECUTION_TRIGGER_INFO = "executionTriggerInfo";
-  private static String TRIGGER_TYPE = "triggerType";
-
-  private static String STEP_ARTIFACTS = "stepArtifacts";
-  private static String PROVENANCE_ARTIFACTS = "provenanceArtifacts";
-  private static String POLICY_OUTPUT = "policyOutput";
-  private static String STATUS = "status";
-
-  private static String PUBLISHED_IMAGES = "publishedImageArtifacts";
-  private static String IMAGE_NAME = "imageName";
-  private static String TAG = "tag";
+  @Inject PipelineService pipelineService;
 
   @Override
   public boolean upsertInstance(Instance instance) {
@@ -92,9 +63,9 @@ public class CdInstanceSummaryServiceImpl implements CdInstanceSummaryService {
         instance.getPrimaryArtifact().getArtifactIdentity().getImage(), instance.getEnvIdentifier());
 
     if (Objects.nonNull(cdInstanceSummary)) {
-      JsonNode rootNode = getPmsExecutionSummary(instance);
+      JsonNode rootNode = pipelineService.getPmsExecutionSummary(instance);
       cdInstanceSummary.getInstanceIds().add(instance.getId());
-      cdInstanceSummary.setSlsaVerificationSummary(getSlsaVerificationSummary(rootNode, instance, artifact));
+      cdInstanceSummary.setSlsaVerificationSummary(pipelineService.getSlsaVerificationSummary(rootNode, artifact));
       cdInstanceSummary = setPipelineDetails(cdInstanceSummary, rootNode, instance);
       cdInstanceSummaryRepo.save(cdInstanceSummary);
     } else {
@@ -214,7 +185,7 @@ public class CdInstanceSummaryServiceImpl implements CdInstanceSummaryService {
 
   @VisibleForTesting
   public CdInstanceSummary createInstanceSummary(Instance instance, ArtifactEntity artifact) {
-    JsonNode rootNode = getPmsExecutionSummary(instance);
+    JsonNode rootNode = pipelineService.getPmsExecutionSummary(instance);
 
     CdInstanceSummary cdInstanceSummary =
         CdInstanceSummary.builder()
@@ -222,7 +193,7 @@ public class CdInstanceSummaryServiceImpl implements CdInstanceSummaryService {
             .accountIdentifier(instance.getAccountIdentifier())
             .orgIdentifier(instance.getOrgIdentifier())
             .projectIdentifier(instance.getProjectIdentifier())
-            .slsaVerificationSummary(getSlsaVerificationSummary(rootNode, instance, artifact))
+            .slsaVerificationSummary(pipelineService.getSlsaVerificationSummary(rootNode, artifact))
             .envIdentifier(instance.getEnvIdentifier())
             .envName(instance.getEnvName())
             .envType(EnvType.valueOf(instance.getEnvType().name()))
@@ -235,110 +206,14 @@ public class CdInstanceSummaryServiceImpl implements CdInstanceSummaryService {
 
   private CdInstanceSummary setPipelineDetails(
       CdInstanceSummary cdInstanceSummary, JsonNode rootNode, Instance instance) {
-    cdInstanceSummary.setLastPipelineName(
-        getNodeValue(parseField(rootNode, PIPELINE_EXECUTION_SUMMARY, PIPELINE_NAME)));
+    cdInstanceSummary.setLastPipelineName(pipelineService.getPipelineName(rootNode));
     cdInstanceSummary.setLastPipelineExecutionName(instance.getLastPipelineExecutionName());
     cdInstanceSummary.setLastPipelineExecutionId(instance.getLastPipelineExecutionId());
-    cdInstanceSummary.setSequenceId(getNodeValue(parseField(rootNode, PIPELINE_EXECUTION_SUMMARY, RUN_SEQUENCE)));
+    cdInstanceSummary.setSequenceId(pipelineService.getPipelineExecutionSequenceId(rootNode));
     cdInstanceSummary.setLastDeployedAt(instance.getLastDeployedAt());
     cdInstanceSummary.setLastDeployedById(instance.getLastDeployedById());
     cdInstanceSummary.setLastDeployedByName(instance.getLastDeployedByName());
-    cdInstanceSummary.setTriggerType(
-        getNodeValue(parseField(rootNode, PIPELINE_EXECUTION_SUMMARY, EXECUTION_TRIGGER_INFO, TRIGGER_TYPE)));
+    cdInstanceSummary.setTriggerType(pipelineService.getPipelineExecutionTriggerType(rootNode));
     return cdInstanceSummary;
-  }
-
-  private SLSAVerificationSummary getSlsaVerificationSummary(
-      JsonNode rootNode, Instance instance, ArtifactEntity artifact) {
-    SLSAVerificationSummaryBuilder slsaVerificationSummaryBuilder = SLSAVerificationSummary.builder();
-    if (rootNode == null) {
-      return slsaVerificationSummaryBuilder.build();
-    }
-    try {
-      JsonNode executionNodeList = parseField(rootNode, EXECUTION_GRAPH, NODE_MAP);
-      if (Objects.isNull(executionNodeList)) {
-        return slsaVerificationSummaryBuilder.build();
-      }
-      for (JsonNode node : executionNodeList) {
-        if (SLSA_VERIFICATION_STEP_ID.equals(getNodeValue(node.get(STEP_TYPE)))
-            && getNodeValue(node.get(IDENTIFIER)) != null) {
-          String fqnSlsaStepIdentifier = getFqnSlsaStepIdentifier(node);
-          if (fqnSlsaStepIdentifier != null && isCorrelated(fqnSlsaStepIdentifier, node, artifact)) {
-            JsonNode provenanceArtifactList =
-                parseField(node, OUTCOMES, fqnSlsaStepIdentifier, STEP_ARTIFACTS, PROVENANCE_ARTIFACTS);
-            JsonNode provenanceArtifact =
-                Objects.nonNull(provenanceArtifactList) ? provenanceArtifactList.get(0) : null;
-            slsaVerificationSummaryBuilder
-                .slsaPolicyOutcomeStatus(getNodeValue(parseField(node, OUTCOMES, POLICY_OUTPUT, STATUS)))
-                .provenanceArtifact(provenanceArtifact);
-            break;
-          }
-        }
-      }
-    } catch (Exception e) {
-      log.error(String.format("Failed to extract SLSA Verification Data. Exception: %s", e));
-    }
-    return slsaVerificationSummaryBuilder.build();
-  }
-
-  private String getFqnSlsaStepIdentifier(JsonNode node) {
-    JsonNode outcomes = parseField(node, OUTCOMES);
-    if (outcomes == null) {
-      return null;
-    }
-    Iterator<String> fields = outcomes.fieldNames();
-    String identifier = getNodeValue(parseField(node, IDENTIFIER));
-    while (fields.hasNext()) {
-      String field = fields.next();
-      if (field.startsWith("artifact") && field.endsWith(identifier)) {
-        return field;
-      }
-    }
-    return null;
-  }
-
-  private boolean isCorrelated(String fqnIdentifier, JsonNode node, ArtifactEntity artifact) {
-    JsonNode publishedImageArtifacts = parseField(node, OUTCOMES, fqnIdentifier, STEP_ARTIFACTS, PUBLISHED_IMAGES);
-    if (publishedImageArtifacts == null) {
-      return false;
-    }
-    for (JsonNode artifactNode : publishedImageArtifacts) {
-      if (artifact.getName().equals(getNodeValue(parseField(artifactNode, IMAGE_NAME)))
-          && artifact.getTag().equals(getNodeValue(parseField(artifactNode, TAG)))) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  private JsonNode getPmsExecutionSummary(Instance instance) {
-    JsonNode rootNode = null;
-    try {
-      Object pmsExecutionSummary = NGRestUtils.getResponse(pipelineServiceClient.getExecutionDetailV2(
-          instance.getLastPipelineExecutionId(), instance.getAccountIdentifier(), instance.getOrgIdentifier(),
-          instance.getProjectIdentifier(), instance.getStageSetupId()));
-      rootNode = JsonUtils.asTree(pmsExecutionSummary);
-    } catch (Exception e) {
-      log.error(String.format("PMS Request Failed. Exception: %s", e));
-    }
-    return rootNode;
-  }
-
-  private JsonNode parseField(JsonNode rootNode, String... path) {
-    for (String field : path) {
-      if (rootNode == null) {
-        return null;
-      } else {
-        rootNode = rootNode.get(field);
-      }
-    }
-    return rootNode;
-  }
-
-  private String getNodeValue(JsonNode node) {
-    if (Objects.isNull(node)) {
-      return null;
-    }
-    return node.asText();
   }
 }
