+++ ssca-manager/service/src/test/java/io/harness/SSCAManagerTestRule.java
@@ -50,6 +50,8 @@ import io.harness.ssca.services.NormalisedSbomComponentService;
 import io.harness.ssca.services.NormalisedSbomComponentServiceImpl;
 import io.harness.ssca.services.OrchestrationStepService;
 import io.harness.ssca.services.OrchestrationStepServiceImpl;
+import io.harness.ssca.services.PipelineService;
+import io.harness.ssca.services.PipelineServiceImpl;
 import io.harness.ssca.services.RuleEngineService;
 import io.harness.ssca.services.RuleEngineServiceImpl;
 import io.harness.ssca.services.S3StoreService;
@@ -185,6 +187,7 @@ public class SSCAManagerTestRule implements InjectorRuleMixin, MethodRule, Mongo
         bind(TokenApi.class).to(TokenApiImpl.class);
         bind(MongoTemplate.class).toInstance(mock(MongoTemplate.class));
         bind(PipelineServiceClient.class).toInstance(mock(PipelineServiceClient.class));
+        bind(PipelineService.class).to(PipelineServiceImpl.class);
       }
     });
     modules.add(TimeModule.getInstance());
