+++ ssca-manager/service/src/test/java/io/harness/ssca/services/ArtifactServiceImplTest.java
@@ -20,6 +20,7 @@ import static org.mockito.Mockito.verify;
 import io.harness.BuilderFactory;
 import io.harness.SSCAManagerTestBase;
 import io.harness.category.element.UnitTests;
+import io.harness.remote.client.NGRestUtils;
 import io.harness.repositories.ArtifactRepository;
 import io.harness.repositories.CdInstanceSummaryRepo;
 import io.harness.repositories.EnforcementSummaryRepo;
@@ -43,10 +44,14 @@ import io.harness.ssca.entities.NormalizedSBOMComponentEntity;
 import io.harness.ssca.entities.NormalizedSBOMComponentEntity.NormalizedSBOMComponentEntityBuilder;
 import io.harness.ssca.utils.PageResponseUtils;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Inject;
+import java.io.IOException;
+import java.io.InputStream;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -57,6 +62,7 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
+import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.springframework.data.domain.Page;
@@ -73,7 +79,6 @@ public class ArtifactServiceImplTest extends SSCAManagerTestBase {
   @Mock ArtifactRepository artifactRepository;
   @Mock EnforcementSummaryRepo enforcementSummaryRepo;
   @Mock CdInstanceSummaryRepo cdInstanceSummaryRepo;
-
   @Inject CdInstanceSummaryService cdInstanceSummaryService;
   private BuilderFactory builderFactory;
 
@@ -203,7 +208,14 @@ public class ArtifactServiceImplTest extends SSCAManagerTestBase {
   @Test
   @Owner(developers = ARPITJ)
   @Category(UnitTests.class)
-  public void testGetArtifactDetails() {
+  public void testGetArtifactDetails() throws IOException {
+    ObjectMapper objectMapper = new ObjectMapper();
+    ClassLoader classLoader = this.getClass().getClassLoader();
+    InputStream inputStream = classLoader.getResourceAsStream("pms-execution-response-with-slsa.json");
+    Map<String, Object> pmsJsonResponse = objectMapper.readValue(inputStream, Map.class);
+    MockedStatic<NGRestUtils> mockRestStatic = Mockito.mockStatic(NGRestUtils.class);
+    mockRestStatic.when(() -> NGRestUtils.getResponse(any())).thenReturn(pmsJsonResponse);
+
     Mockito.when(artifactRepository.findOne(any())).thenReturn(builderFactory.getArtifactEntityBuilder().build());
 
     ArtifactDetailResponse response = artifactService.getArtifactDetails(builderFactory.getContext().getAccountId(),
@@ -217,6 +229,7 @@ public class ArtifactServiceImplTest extends SSCAManagerTestBase {
     assertThat(response.getNonProdEnvCount()).isEqualTo(1);
     assertThat(response.getBuildPipelineId()).isEqualTo("pipelineId");
     assertThat(response.getBuildPipelineExecutionId()).isEqualTo("pipelineExecutionId");
+    assertThat(response.getBuildPipelineName()).isEqualTo("SLSA attestation and verification");
   }
 
   @Test
